4.4.0 (#156)
* Update CHANGELOG.md

* Update package.json

* Added humStepSize

* humStepSize

* humStepSize

* Update humidifier-dehumidifier.js

* Added KeepAlives

Fixes #113

* fixed broadcast error

* Added unlocking details

* Typo fix

* Update CHANGELOG.md

* Update humidifier-dehumidifier.js

* Update CHANGELOG.md

* Update README.md

* Added 50 humidity default when off

* Fixed "no response" error with noHumidity

Fixes #116

* Update CHANGELOG.md

* Update package.json

* Fix for log not found error using Pronto

* Update CHANGELOG.md

* Update package.json

* Update package.json

* Added "disabled" accessories to remove temporarily

* Support for hum and temp in Files

* Added hum and Temp from file

* Update humidifier-dehumidifier.js

* Update aircon.js

* Update aircon.js

* Fixed dupliate MonitorTemperature calls

* Update CHANGELOG.md

* Update CHANGELOG.md

* Added Battery Alerts

* Added Battery Alerts

* Added battery level from file

* Added battery Alert support

* Update CHANGELOG.md

* Update package.json

* Update humiditySensor.js

* Update temperatureSensor.js

* Added Battery Level

* Added Battery Level

* Added Battery and File details

* Removed duplicate SetDefaults call

* Added MQTT Humidity,Battery

* Improved logging

* Fixed name not defined error

* Fixed name not defined error

* Moved to fixed homebridge-platform-helper-kc

* Update platform.js

* Update accessory.js

* Added MQTT details

* Update CHANGELOG.md

* Added battery mqtt support

* Fixed noHumidity support

* Updated homebridge-platform-helper-kc version

1.1.0 removes vulnerabilities

* Update package.json

* Reverted f33b97c

* MQTT value handling cleanup

* MQTT value handling cleanup

* MQTT value handling cleanup

* Update package.json (#145)

* Update package.json

* Update CHANGELOG.md

* Fixed typo

* Typo 2

* Update humidifier-dehumidifier.js

* Update humidifier-dehumidifier.js

* Added MQTT data and moved base documentation

Moved base documentation to top to assist new users

* Update CHANGELOG.md

* Spelling

* Support for heater-cooler accessory with rotation speed and swing mode (#147)

* Update CHANGELOG.md

* Update package.json

* Integrate platform helper (#152)

Thanks @newt10

* Update CHANGELOG.md

* Increased MQTT default temperatureUpdateFrequency 

MQTT updates when published so frequent refreshes aren't required ( 10 minute default as a fallback )

* Update CHANGELOG.md

* Update package.json

* Increased RF Scan timeout to 30 seconds

Increased to match Broadlink app behaviour

* Updated documentation link

* Updated URL and removed duplicated info

* Fix dependancy versions

* Bump bl from 1.2.2 to 4.0.3 (#155)

* Bump bl from 1.2.2 to 4.0.3

Bumps [bl](https://github.com/rvagg/bl) from 1.2.2 to 4.0.3.
- [Release notes](https://github.com/rvagg/bl/releases)
- [Commits](https://github.com/rvagg/bl/compare/v1.2.2...v4.0.3)

+++ platform.js
@@ -1,4 +1,4 @@
-const { HomebridgePlatform } = require('homebridge-platform-helper');
+const { HomebridgePlatform } = require('./base');
 const { assert } = require('chai');
 
 const npmPackage = require('./package.json');
@@ -28,7 +28,8 @@ const classTypes = {
   'window-covering': Accessory.WindowCovering,
   'tv': Accessory.TV,
   'temperatureSensor': Accessory.TemperatureSensor,
-  'humiditySensor': Accessory.HumiditySensor
+  'humiditySensor': Accessory.HumiditySensor,
+  'heater-cooler': Accessory.HeaterCooler
 }
 
 let homebridgeRef
@@ -69,7 +70,7 @@ const BroadlinkRMPlatform = class extends HomebridgePlatform {
     let tvs = [];
     config.accessories.forEach((accessory) => {
       if (!accessory.type) throw new Error(`Each accessory must be configured with a "type". e.g. "switch"`);
-
+      if (accessory.disabled) return;
       if (!classTypes[accessory.type]) throw new Error(`homebridge-broadlink-rm doesn't support accessories of type "${accessory.type}".`);
 
       const homeKitAccessory = new classTypes[accessory.type](log, accessory);
